+++ compiler/lic/main.cpp
@@ -13,6 +13,7 @@
 #include "li/terminal.hpp"
 #include "li/error_codes.hpp"
 #include "li/proof_cli.hpp"
+#include "li/resource_options.hpp"
 
 #include "li_rt.h"
 
@@ -156,16 +157,17 @@ int httpd_explain_config(int argc, char** argv) {
   return 0;
 }
 
-bool apply_resource_flag(std::string_view arg) {
-  if (arg.rfind("--jobs=", 0) == 0) {
-    setenv("LI_COMPILE_JOBS", std::string(arg.substr(7)).c_str(), 1);
-    return true;
+void apply_resource_options_to_env() {
+  const auto& opts = li::resource_options();
+  if (!opts.build_dir.empty()) {
+    setenv("LI_BUILD_DIR", opts.build_dir.c_str(), 1);
   }
-  if (arg.rfind("--max-memory=", 0) == 0) {
-    setenv("LI_MAX_MEMORY_MB", std::string(arg.substr(13)).c_str(), 1);
-    return true;
+  if (opts.threads > 0) {
+    setenv("LI_OMP_THREADS", std::to_string(opts.threads).c_str(), 1);
   }
-  return false;
+  if (opts.max_memory_mb > 0) {
+    setenv("LI_MAX_MEMORY_MB", std::to_string(opts.max_memory_mb).c_str(), 1);
+  }
 }
 
 std::string read_file(const char* path) {
@@ -403,6 +405,7 @@ int build_file(const char* path, const char* output, const li::CompileOptions& o
 }  // namespace
 
 int main(int argc, char** argv) {
+  li::reset_resource_options();
   if (argc < 2) {
     return usage();
   }
@@ -478,10 +481,14 @@ int main(int argc, char** argv) {
       } else if (std::string_view(argv[i]) == "--strict-lean") {
         run_lean = true;
         strict_lean = true;
+      } else if (li::apply_resource_flag(argv[i], li::resource_options())) {
+        continue;
       } else if (parse_proof_cli_flag(argv[i])) {
         continue;
       }
     }
+    li::finalize_resource_options(li::resource_options());
+    apply_resource_options_to_env();
     return verify_file(argv[2], run_lean, strict_lean);
   }
   if (cmd == "validate-httpd-config") {
@@ -529,15 +536,16 @@ int main(int argc, char** argv) {
         strict_lean = true;
       } else if (parse_proof_cli_flag(arg)) {
         continue;
-      } else if (arg.rfind("--threads=", 0) == 0) {
-        setenv("LI_OMP_THREADS", std::string(arg.substr(10)).c_str(), 1);
-      } else if (apply_resource_flag(arg)) {
+      } else if (li::apply_resource_flag(arg, li::resource_options())) {
         continue;
       } else {
         extra_flags.append(argv[i]);
         extra_flags.push_back(' ');
       }
     }
+    li::finalize_resource_options(li::resource_options());
+    li::note_compile_jobs_reserved(li::resource_options());
+    apply_resource_options_to_env();
     if (coverage) {
       extra_flags += "-fprofile-instr-generate -fcoverage-mapping ";
     }